requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
requests==2.32.3
numpy==2.1.3
pandas==2.2.3
pyproj==3.6.1
shapely==2.0.6
supabase==2.9.1
python-dotenv==1.0.1
# matplotlib and reportlab removed (PDF report feature dropped)
